+++ example/example.dart
@@ -1,3 +1,0 @@
-void main() {
-  // For particular examples, visit the files in the `example/` directory.
-}