+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.4.0
+
+- Added `intl` as a dependency.
+- Added `IntlQuantityDisplayMode`.
+- Renamed `SimpleDisplayMode` to `SimpleQuantityDisplayMode` for consistency.
+- Deprecated `CustomQuantityDisplayMode` in favour of concrete implementations of `QuantityDisplayMode`.
+
 ## 0.3.0
 
 - Added method `bytes` to `Unit` for more natural comparisons.